[Fixed #438] *: fix bug (#440)
<!-- Thank you for contributing to TiDB!

PR Title Format:
1. [#issue-number] pkg [, pkg2, pkg3]: what's changed
2. [#issue-number] *: what's changed

Please fill in the number of the issue resolved by the PR in the title, if not, please add `issue-number-none`.

-->

### What problem does this PR solve?

Issue Number: close #438 <!-- REMOVE this line if no issue to close -->

Problem Summary: a danger bug.

### What is changed and how it works?

Proposal: [xxx](url) <!-- REMOVE this line if not applicable -->

What's Changed: a little change.

How it Works:

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [x] No code

Side effects

- [] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [x] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note <!-- bugfixes or new feature need a release note -->

```release-note
None
```

<!-- Thank you for contributing to TiDB!

PR Title Format:
1. [#issue-number] pkg [, pkg2, pkg3]: what's changed
2. [#issue-number] *: what's changed

Please fill in the number of the issue resolved by the PR in the title, if not, please add `issue-number-none`.

-->

### What problem does this PR solve?

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary:

### What is changed and how it works?

Proposal: [xxx](url) <!-- REMOVE this line if not applicable -->

What's Changed:

How it Works:

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note <!-- bugfixes or new feature need a release note -->

```release-note
Please add a release note, or a 'None' if it is not needed.
```